[Decomp][Object] Cohesion split: extract summon helpers to WorldObjectSummon.cpp
